Avoid overwriting the first typeahead find letter
Grabbing the focus of the search entry caused the whole contents of the
entry to be selected. As this was done when entering a character to
bring up the search entry, it was easy to type a whole word, overwriting
the first (selected) character.

Fix the problem by setting the cursor position to the end of the search
entry immediately after grabbing focus.