journal: allow matching on field existence
The event view had to manually filter results that don't have a
_KERNEL_DEVICE or _AUDIT_SESSION (for hardware and security,
respectively). Move this logic into the GlJournal, where all other
matching is done. To match on existence, pass only a FIELDNAME instead
of a FIELDNAME=VALUE pair in the query.

This needs to be emulated in GlJournal, because systemd's journal API
doesn't support it.