journalmodel: load entries in batches
A major use case of gnome-logs is to open it to quickly see the most
recent log entries. It doesn't make sense to read all available entries
from disk and create row widgets for each of them.

Thus, make GlJournalModel only load the first 50 entries and trigger it
to load more once the user scrolls the listbox to the bottom.

We still have to load all entries once a search is started.

https://bugzilla.gnome.org/show_bug.cgi?id=728317