Make comments to be seen by translators
The problem is that there are comments for translators which are two
lines above the gettext call _("...") and these will not be seen by
translators.

To fix this rearrange lines in order to place the commentaries right
after the parameter. Also keep the coding style by indenting the
comments and having no more than 79 characters on a line.

https://bugzilla.gnome.org/show_bug.cgi?id=763572