Add search popover with journal field filtering
The search popover can be selected from the
drop down menu besides the search bar. Individual
journal field or all available journal fields can
be selected from the search popover.The default
journal field filter is "All Available fields".

The search popover is designed to look similar to the
Nautilus search popover.

If Nautilus implementation of search popover changes in
future, gnome-logs implementation of search popover should
be kept in sync with it.

https://bugzilla.gnome.org/show_bug.cgi?id=767996