Add custom journal timestamp range submenu
The custom timestamp range submenu is opened through the select
timestamp range treeview.

If the user doesn't specify any date, then today's date is taken as
default. If the user doesn't specify any time, then 11:59:59 PM is
taken as default start time and 12:00:00 AM is taken default end time.
This is done so that if the user selects the same date in both the
start and end date selection widgets, all the logs corresponding to
that date will be shown in the window.

The journal is filtered according to the entered start date-time and
end date-time timestamp range and the logs are shown in window
accordingly.

Invalid values in time spinboxes are set to nearest correct values.
Future dates fail silently.

If the user selects any other filter, then the options
in the custom timestamp range submenu are reset.

Depending upon whether the user has selected start date-time
or end date-time, we update the timestamp range button label
accordingly.

If time zone format is 24hr, we show only the hour-minute-second
spinboxes in the custom timestamp range submenu.

https://bugzilla.gnome.org/show_bug.cgi?id=767996