Compress similar events in GlEventViewList
The compression information is stored in GlRowEntry objects
during population of the model array.The similarity criteria
for grouping adjacent messages is:

* Messages whose first word is same
* Messages sent by the same process

The compression information is then parsed in the view classes
and accordingly the events to be compressed are hidden in GlEventViewList.

The visibility of compressed events can be toggled by
activating the header row representing those compressed events.

When compressed events are expanded using the header, sorting order
specified by 'sort-order' GSettings key is maintained in GlEventViewList.

Headers are removed from the compressed rows to improve their visibility
when they are shown by activating the header.

https://bugzilla.gnome.org/show_bug.cgi?id=709294