Display detailed event information in a GtkPopover
A GtkPopover is shown when a row which is not a header is
activated. For handling very long messages, the message text is
wrapped to fit the size of the GtkPopover. Further, such long
messages are shown in a GtkScrolledWindow. All the earlier
available detail fields are shown in the GtkPopover.

https://bugzilla.gnome.org/show_bug.cgi?id=709294