Allow users to ignore insufficient permission warning
Currently, if an user don't have sufficient permission to view logs,
a GtkInfoBar will show up to warn the user about it. And it shows up
every time the user opens Logs until the user has sufficient
permission. This can be quite annoying.

Fix this by adding an "ignore" button to GtkInfoBar to give users the
option to ignore the warning.