update boot timestamps when possible
Previously, boot timestamps in the boot selection menu is generated
when Logs is opened and won't be updated later. This causes a problem
that the timestamp of the latest log entry might be newer than what's
shown in the boot selection menu.

Fix that by updating the boot selection menu when users change the
category list or view logs from other boots.

https://bugzilla.gnome.org/show_bug.cgi?id=762840