Consume the new entries, and add them to the model
Receive the signal to consume the new entries, the approach to
adding them to the model seems similar to the function
gl_journal_model_fetch_idle (), so factor out the common code to
be a separate function gl_journal_model_process_add ().

The function gl_journal_model_add_header () was modified to be
reused for adding the new headers.