Add flatpak manifest
This is useful both for development[0] and for distribution in the
gnome-apps-nightly repository. It is also a good base for a future
stable manifest on flathub.

Unfortunately the manifest isn't as simple as hoped: The current
platform contains a libsystemd without lz4 support, which means
that it cannot open any journal files if the host stores them with
the preferred compression. That is, to be actually useful on common
systems, we have to rebuild libsystemd ourselves ...

[0] https://csorianognome.wordpress.com/2017/04/07/the-new-contribution-workflow-for-gnome/

https://gitlab.gnome.org/GNOME/gnome-logs/issues/39