searchpopover: Replace GtkTreeView with GtkDropDown
Drop a bunch of code related to the TreeView and all of the stacks,
etc that are necessar to make TreeView work as a drop down.  Use
sections in the dropbox to replace the separators that previously
existed in the treeview.  There's a tiny bit of fragility to the
order of the dropdown, but the complexity in using a custom GObject,
doesn't seem with it.

Related: https://gitlab.gnome.org/GNOME/Initiatives/-/issues/49